requirements.txt
Flask==2.3.3
rdkit-pypi==2022.9.5
